Refactor project structure to organize under 'civis' package
- Moved all source files and servers into 'civis/' directory for better package structure.
- Updated imports throughout the codebase to reflect new paths.
- Added `civis/__init__.py` to initialize the package with versioning and module imports.
- Updated `.gitignore` to exclude `civis/config.json` and `civis.egg-info/` directories.
- Refactored `app.py` to use command-line arguments as a global `args` variable.
- Added a `setup.py` file for package distribution.
