+++ civis/servers/trajectory_servers/trajectory_server_v6.py
@@ -4,8 +4,10 @@ import os
 import sys
 
 from bokeh.plotting import figure, curdoc
-from bokeh.models import ColumnDataSource, Slider, Button, Arrow, VeeHead, Div, TextInput, Spacer
+from bokeh.models import ColumnDataSource, Slider, Button, Arrow, VeeHead, Div, TextInput, Spacer, LinearColorMapper, \
+    ColorBar, BasicTicker, Select
 from bokeh.layouts import column, row
+from bokeh.palettes import Blues8
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 servers_dir = os.path.dirname(current_dir)
@@ -13,9 +15,9 @@ project_root = os.path.dirname(servers_dir)
 sys.path.append(project_root)
 
 
-def trajectory_bkapp_v7(doc):
-    from src import VirmenTank
-    global source, trials
+def trajectory_bkapp_v6(doc):
+    from civis.src.VirmenTank import VirmenTank
+    global source, trials, confusion_matrix_source, correct_array
 
     trials = []
     source = ColumnDataSource({'x': [], 'y': [], 'face_angle': []})
@@ -49,16 +51,51 @@ def trajectory_bkapp_v7(doc):
     previous_button = Button(label="Previous", width=100)
     next_button = Button(label="Next", width=100)
     starts_div = Div(text="Start Time: ", width=400)
+    correctness_div = Div(text="Trial Correctness: ", width=400)
     error_div = Div(text="")
+
+    # New dropdown menus for correct and incorrect trials
+    correct_trials_dropdown = Select(title="Correct Trials:", value="", options=[], width=200)
+    incorrect_trials_dropdown = Select(title="Incorrect Trials:", value="", options=[], width=200)
 
     trial_slider.disabled = True
     progress_slider.disabled = True
     play_button.disabled = True
     previous_button.disabled = True
     next_button.disabled = True
+    correct_trials_dropdown.disabled = True
+    incorrect_trials_dropdown.disabled = True
+
+    # Initialize confusion matrix plot with default data
+    default_matrix = np.array([[-1, -1], [-1, -1]])
+    x_labels = ['Left', 'Right']
+    y_labels = ['Left', 'Right']
+    confusion_matrix_source = ColumnDataSource(data={
+        'x': [x for x in x_labels for _ in y_labels],
+        'y': y_labels * len(x_labels),
+        'value': default_matrix.flatten().tolist(),
+    })
+
+    color_mapper = LinearColorMapper(palette=Blues8[::-1], low=-1, high=1)
+
+    confusion_matrix_plot = figure(title="Confusion Matrix of Animal Turns",
+                                   x_range=x_labels, y_range=list(reversed(y_labels)),
+                                   y_axis_label="Actual Turn", x_axis_label="Should Turn",
+                                   width=400, height=350, toolbar_location=None, tools="")
+
+    confusion_matrix_plot.rect(x='x', y='y', width=1, height=1, source=confusion_matrix_source,
+                               line_color=None, fill_color={'field': 'value', 'transform': color_mapper})
+
+    color_bar = ColorBar(color_mapper=color_mapper, ticker=BasicTicker(desired_num_ticks=len(Blues8)),
+                         label_standoff=6, border_line_color=None, location=(0, 0))
+
+    confusion_matrix_plot.add_layout(color_bar, 'right')
+
+    confusion_matrix_plot.text(x='x', y='y', text='value', source=confusion_matrix_source,
+                               text_align="center", text_baseline="middle")
 
     def load_data():
-        global source, trials, starts
+        global source, trials, starts, confusion_matrix_source, correct_array
 
         try:
             config_path = os.path.join(project_root, 'config.json')
@@ -72,14 +109,15 @@ def trajectory_bkapp_v7(doc):
                 file = os.path.join(config['VirmenFilePath'], f'{session_name}.txt')
 
             # Check if the maze type is correct
-            if VirmenTank.determine_maze_type(file).lower() != "turnv0":
+            if VirmenTank.determine_maze_type(file).lower() != "turnv1":
                 raise ValueError(
-                    f"Invalid maze type. Expected 'TurnV0', but got '{VirmenTank.determine_maze_type(file)}'")
+                    f"Invalid maze type. Expected 'TurnV1', but got '{VirmenTank.determine_maze_type(file)}'")
             else:
                 vm = VirmenTank(file, height=35)
 
             trials = vm.virmen_trials
             starts = [x / vm.vm_rate for x in vm.trials_start_indices]
+            correct_array = vm.extend_data.correct_array
 
             print("Successfully loaded: " + file)
             error_div.text = ""  # Clear any previous error messages
@@ -91,6 +129,8 @@ def trajectory_bkapp_v7(doc):
                 play_button.disabled = False
                 previous_button.disabled = False
                 next_button.disabled = False
+                correct_trials_dropdown.disabled = False
+                incorrect_trials_dropdown.disabled = False
 
                 initial_trial = trials[0]
                 new_data = {'x': [initial_trial['x'][0]],
@@ -106,6 +146,31 @@ def trajectory_bkapp_v7(doc):
 
                 starts_div.text = f"Start Time: {starts[0]}"
 
+                if correct_array[0]:
+                    correctness_div.text = f"Trial Correctness: Correct"
+                else:
+                    correctness_div.text = f"Trial Correctness: Wrong"
+
+                # Update confusion matrix plot
+                new_confusion_matrix = vm.extend_data.confusion_matrix
+                confusion_matrix_source.data.update({
+                    'value': new_confusion_matrix.flatten().tolist()
+                })
+                color_mapper.low = new_confusion_matrix.min()
+                color_mapper.high = new_confusion_matrix.max()
+
+                # Update dropdown menus
+                correct_trials = [f"Trial {i}" for i, correct in enumerate(correct_array) if correct]
+                incorrect_trials = [f"Trial {i}" for i, correct in enumerate(correct_array) if not correct]
+
+                correct_trials_dropdown.options = correct_trials
+                incorrect_trials_dropdown.options = incorrect_trials
+
+                if correct_trials:
+                    correct_trials_dropdown.value = correct_trials[0]
+                if incorrect_trials:
+                    incorrect_trials_dropdown.value = incorrect_trials[0]
+
         except ValueError as e:
             error_div.text = f"Error: {str(e)}"
             # Disable widgets if data loading fails
@@ -114,6 +179,8 @@ def trajectory_bkapp_v7(doc):
             play_button.disabled = True
             previous_button.disabled = True
             next_button.disabled = True
+            correct_trials_dropdown.disabled = True
+            incorrect_trials_dropdown.disabled = True
         except Exception as e:
             error_div.text = f"An unexpected error occurred: {str(e)}"
             # Disable widgets if data loading fails
@@ -122,6 +189,8 @@ def trajectory_bkapp_v7(doc):
             play_button.disabled = True
             previous_button.disabled = True
             next_button.disabled = True
+            correct_trials_dropdown.disabled = True
+            incorrect_trials_dropdown.disabled = True
 
     load_button.on_click(load_data)
 
@@ -152,6 +221,11 @@ def trajectory_bkapp_v7(doc):
                     'face_angle': trial_data['face_angle'][:max_index]}
         source.data = new_data
         starts_div.text = f"Start Time: {starts[trial_index]}"
+
+        if correct_array[trial_index]:
+            correctness_div.text = f"Trial Correctness: Correct"
+        else:
+            correctness_div.text = f"Trial Correctness: Wrong"
 
     trial_slider.on_change('value', update_plot)
     progress_slider.on_change('value', update_plot)
@@ -200,13 +274,22 @@ def trajectory_bkapp_v7(doc):
     previous_button.on_click(previous_trial)
     next_button.on_click(next_trial)
 
+    def update_trial_from_dropdown(attr, old, new):
+        if new:
+            trial_index = int(new.split()[1])
+            trial_slider.value = trial_index
+
+    correct_trials_dropdown.on_change('value', update_trial_from_dropdown)
+    incorrect_trials_dropdown.on_change('value', update_trial_from_dropdown)
+
     file_input_row = row(filename_input, column(Spacer(height=20), load_button))
     trial_navigation_row = row(previous_button, next_button)
+    dropdown_row = row(correct_trials_dropdown, incorrect_trials_dropdown)
     tool_widgets = column(file_input_row, trial_slider, progress_slider, play_button,
-                          trial_navigation_row, starts_div, error_div)
-    layout = row(plot, Spacer(width=30), tool_widgets)
+                          trial_navigation_row, starts_div, correctness_div, dropdown_row, error_div)
+    layout = row(plot, Spacer(width=30), column(tool_widgets, confusion_matrix_plot))
     doc.add_root(layout)
 
 
 # Uncomment the following line to run the Bokeh server application
-trajectory_bkapp_v7(curdoc())
+trajectory_bkapp_v6(curdoc())
