+++ civis/templates/labeler.html
