+++ civis/templates/place/place_base.html
