+++ civis/templates/place/place_v2.html
