+++ civis/templates/place/place_v3.html
