+++ civis/templates/raster/raster_base.html
