+++ civis/templates/trajectory/trajectory_base.html
