--- src/__init__.py
@@ -1,3 +1,0 @@
-from src.CITank import CITank
-from src.VirmenTank import VirmenTank
-from src.ElecTank import ElecTank
